Hallelujah the rendering is fixed, all hail compiler warnings

+++ linuxdoom-1.10/r_bsp.cxx
@@ -303,7 +303,7 @@ void R_AddLine (seg_t*	line)
 	// Totally off the left edge?
 	if (tspan >= span)
 	    return;	
-	angle2 = -clipangle;
+	angle2 = -static_cast<int32_t>(clipangle);
     }
     
     // The seg is in the view range,
@@ -455,7 +455,7 @@ boolean R_CheckBBox (fixed_t*	bspcoord)
 	if (tspan >= span)
 	    return false;
 	
-	angle2 = -clipangle;
+	angle2 = -static_cast<int32_t>(clipangle);
     }
